+++ flashback/debugging/__init__.py
@@ -1,8 +1,10 @@
+from .get_frame import get_frame
 from .profiled import profiled
 from .xp import xp
 
 
 __all__ = (
+    'get_frame',
     'profiled',
     'xp',
 )